requirements-dev.txt
# Development dependencies for MyAIStudio Backend
# Install with: pip install -r requirements-dev.txt

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code formatting and linting
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Development tools
python-dotenv==1.0.0
watchdog==3.0.0

# Database tools
alembic==1.12.1

# API documentation
fastapi==0.104.1
uvicorn[standard]==0.24.0





















